fix createUser, chức năng uploadImages có thể tải lên nhiều ảnh cùng lúc

+++ src/main/java/com/example/AuctionSite/service/ImageService.java
@@ -17,6 +17,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.List;
 
 @Slf4j
@@ -29,24 +30,35 @@ public class ImageService {
     
     String UPLOAD_DIR = "src/main/resources/static/images/";
     
-    @PreAuthorize("hasAuthority('UPLOAD_IMAGE')")
-    public ImageResponse uploadImage(ImageRequest imageRequest) throws IOException {
-        MultipartFile file = imageRequest.getFile();
+    @PreAuthorize("hasAuthority('UPLOAD_IMAGES')")
+    public List<ImageResponse> uploadImages(ImageRequest imageRequest) throws IOException {
+        List<ImageResponse> imageResponses = new ArrayList<>();
         
-        if (file.isEmpty()) {
-            throw new IllegalArgumentException("File is empty");
+        for (MultipartFile file : imageRequest.getFiles()) {
+            if (file.isEmpty()) {
+                throw new IllegalArgumentException("File is empty");
+            }
+            
+            String fileName = file.getOriginalFilename();
+            String filePath = UPLOAD_DIR + fileName;
+            Path path = Paths.get(filePath);
+            
+            Files.createDirectories(path.getParent());
+            
+            if (imageRepository.existsByImageURL("/images/" + fileName)) {
+                throw new RuntimeException("Image existed: " + fileName);
+            }
+            
+            Files.write(path, file.getBytes());
+            
+            Image image = Image.builder()
+                .imageURL("/images/" + fileName)
+                .build();
+            imageResponses.add(imageMapper.toImageResponse(imageRepository.save(image)));
         }
-        
-        String filePath = UPLOAD_DIR +file.getOriginalFilename();
-        Path path = Paths.get(filePath);
-        Files.createDirectories(path.getParent());
-        Files.write(path, file.getBytes());
-        
-        Image image = Image.builder()
-            .imageURL("/images/" + file.getOriginalFilename())
-            .build();
-        return imageMapper.toImageResponse(imageRepository.save(image));
+        return imageResponses;
     }
+    
     
     @PreAuthorize("hasAuthority('GET_ALL_IMAGES')")
     public List<ImageResponse> getAllImages() {
@@ -61,26 +73,38 @@ public class ImageService {
     
     @PreAuthorize("hasAuthority('UPDATE_IMAGE')")
     public ImageResponse updateImage(Integer id, ImageRequest imageRequest) throws IOException {
+        List<MultipartFile> files = imageRequest.getFiles();
+        if (files.size() != 1) {
+            throw new RuntimeException("Only one image can be uploaded at a time");
+        }
+        
         Image image = imageRepository.findById(id).orElseThrow(() -> new RuntimeException("Image not found"));
+        
+        MultipartFile newFile = files.getFirst();
+        String newFileName=newFile.getOriginalFilename();
+        String newFilePath = UPLOAD_DIR + newFileName;
+        Path path = Paths.get(newFilePath);
+        
+        Files.createDirectories(path.getParent());
+        
+        if (imageRepository.existsByImageURL("/images/" + newFileName)) {
+            throw new RuntimeException("Image existed: " + newFileName);
+        }
         
         String oldFilePath = "src/main/resources/static" + image.getImageURL();
         Files.deleteIfExists(Paths.get(oldFilePath));
         
-        MultipartFile newFile = imageRequest.getFile();
-        String newFilePath = UPLOAD_DIR +newFile.getOriginalFilename();
-        Path path = Paths.get(newFilePath);
-        Files.createDirectories(path.getParent());
         Files.write(path, newFile.getBytes());
         
-        image.setImageURL("/images/" +newFile.getOriginalFilename());
+        image.setImageURL("/images/" + newFile.getOriginalFilename());
         return imageMapper.toImageResponse(imageRepository.save(image));
     }
     
     @PreAuthorize("hasAuthority('DELETE_IMAGE')")
     public void deleteImage(Integer id) throws IOException {
         Image image = imageRepository.findById(id).orElseThrow(() -> new RuntimeException("Image not found"));
-        String oldFilePath = "src/main/resources/static" + image.getImageURL();
-        Files.deleteIfExists(Paths.get(oldFilePath));
+        String filePath = "src/main/resources/static" + image.getImageURL();
+        Files.deleteIfExists(Paths.get(filePath));
         imageRepository.deleteById(id);
     }
 }
